fix: modular page
